Create ConsecutiveSums.java
Inspired by Parker, Matt. Totally addicted to base. In: Things to make and do in the fourth dimension. Strauss and Giroux Publishing. 2014. 

A number is said to be a consequential number (own terminology) if it can be composed of sums of consecutive integers; for example:
* 9 = 4 + 5
* 11 = 5 + 6
* 12 = 3 + 4 + 5

Using this concept of consequential numbers, I tried to find how to compose any number N using the largest string of consecutive numbers. 

Discoveries:
* All numbers who cannot be composed this way appear to be powers of 2. 
* All numbers who need a 0 at the start of the chain to be composed of the largest amount of consecutive integers: 6, 10, 28, 136, 496...
* 50 = sum of all numbers from [-10, 25].
* 100 = sum of all numbers from [-17, 40].
* 500 = sum of all numbers from [-97, 200].
* 1 000 = sum of all numbers from [-197, 400]. Appears strangely related to the N = 500 case.